Moved graphics parts to internal directory
