+++ graphics/texture.go
@@ -1,7 +1,6 @@
 package graphics
 
 import (
-	"github.com/GomeBox/gome/adapters/graphics"
 	"github.com/GomeBox/gome/primitives"
 )
 
@@ -9,34 +8,4 @@ type Texture interface {
 	Draw(source, dest *primitives.Rectangle) error
 	DrawF(source *primitives.Rectangle, dest *primitives.RectangleF) error
 	Dimensions() primitives.Dimensions
-}
-
-func newTexture(tex graphics.Texture) Texture {
-	texture := new(texture)
-	texture.tex = tex
-	return texture
-}
-
-type texture struct {
-	tex graphics.Texture
-}
-
-func (texture *texture) Draw(source, dest *primitives.Rectangle) error {
-	err := texture.tex.Draw(source, dest)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (texture *texture) DrawF(source *primitives.Rectangle, dest *primitives.RectangleF) error {
-	err := texture.tex.DrawF(source, dest)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (texture *texture) Dimensions() primitives.Dimensions {
-	return texture.tex.Dimensions()
 }
